fix: set postcss.options.cssnano.colormin to false
Since on development, the sass file will pass colormin, its color may be transfrom by cssnano. In that case, some colors in the color scheme won't be accessible by ui itself. But it can be turned on in production mode, because at that time the color scheme file won't pass the sass loader but a naive rollup plugin, the color won't be changed anymore.